refactoring: rename PhonesContext to AAA
